--- src/main/java/org/text/processor/action/interpretator/Factory.java
@@ -1,9 +1,0 @@
-package org.text.processor.action.interpretator;
-
-public interface Factory {
-    Expression createBitwiseExpression(BitwiseOperator operator, Expression left, Expression right);
-
-    Expression createNotExpression(Expression operand);
-
-    Expression createNumberExpression(int operand);
-}
